requirements.txt
Flask==3.0.2
qrcode==8.0
pillow==10.2.0
